requirements.txt
# Existing requirements
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain>=0.3.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.0.0
aiofiles>=23.0.0

# Vector Databases
chromadb>=0.4.0
pymilvus>=2.3.0  # ← Add Milvus support
sentence-transformers>=2.2.0
openai>=1.0.0

# Streamlit
streamlit>=1.28.0
streamlit-extras>=0.3.0

# Utilities
python-dotenv>=1.0.0
asyncio>=3.4.3
uuid>=1.30
numpy>=1.24.0